+++ packages/app-deploy-ops/rollup/tools/prod-index-plugin.js
@@ -1,39 +1,30 @@
 /*
-* tools/prod-index-filter.js
+* rollup/tools/prod-index-filter.js
 *
-* Only applies to ROLLUP production builds; not Vite.
-*
-* Provide a production version of 'index.html', based on comments within it.
+* Creates a production version of 'index.html', based on comments within a template.
 *
 * Index syntax:
 *   - leave everything, except:
 *
-*     <!--DEV ...-->     // such a block is for development and gets removed
-*     ...
-*     <!-- -->
+*     ${PRELOADS}       // replaced by a modulepreloading code for the chunks produced by Rollup
 *
-*     <!--[DEV,]PROD[,] ...   // such a block is for production. It gets uncommented. Whole first line gets removed (commentary).
-*     ...
-*     -->
-*
-*     ${PRELOADS}       // gets replaced by a modulepreloading code for the chunks produced by Rollup
-*
-*     ${VERSION}        // gets replaced by the 'version' from 'package.json'
+*     ${VERSION}        // replaced by the 'version' from 'package.json'
 *
 *     <script ... src=".../mod.#..."         // '#' gets replaced by the hash for 'mod'
 */
 import { strict as assert } from 'assert'
-import fs from "fs"
+import fs from 'fs'
 
 /*
 * Create a block of:
 *   <<
-*     <link rel="modulepreload" href="dist/@vue-6ed1e3f4.js">
+*     <link rel="modulepreload" href="dist/@vue-6ed1e3f4.js">     <-- tbd. update with non-vue sample
 *     ...
 *   <<
 *
-* Note: 'main' chunk is NOT preloaded. Doing so would allow the browser to compile (and likely run the root source),
+* Note: 'app' chunk is NOT preloaded. Doing so would allow the browser to compile (and likely run the root source),
 *     when we wish to dynamically load it in. Also, we refer to it from HTML anyways, so there is no speedup increase.
+*     tbd. ^-- review comment
 */
 function preloadsArr(modFiles) {   // array of "dist/{mod}-{hash}.js" -> array of "<link rel...>"
 
@@ -50,69 +41,51 @@ function preloadsArr(modFiles) {   // array of "dist/{mod}-{hash}.js" -> array o
 function productize(contents, hashes, { version }) {    // (String, Map<String,String>, { version: String }) -> String    // may throw
   assert(version, "No '.version'");
 
-  const errors = new Array();
-
   const modFiles = Array.from( hashes.entries() ).map( ([key,value]) => `dist/${key}-${value}.js` );
 
   const preloads = preloadsArr(modFiles);   // ["<link ...>", ...]
 
   // Regex note: '.' for "chars within the same line" and '[\s\S]' for "chars, possibly including newline".
 
-  // DEV blocks (remove)
+  // The rest are per-line replacements.
   //
-  const s1 = contents.replace(/^\s*<!--DEV\s.*-->[\s\S]+?^\s*<!--\s+-->\s*\n/gm,"");  // "<!-- DEV block REMOVED -->\n"
+  const outArr = contents.split('\n').map( s => {
+    // '${PRELOADS}'
+    const s2 = s.replace(/^(\s*)\${PRELOADS}\s*$/, (_,c1) => preloads.map(x => c1+x).join('\n') );
+    if (s2 !== s) return s2;    // note: may (does) contain newlines
 
-  // DEV,PROD blocks (keep the inside)
-  //
-  const s2 = s1.replace(/^\s*<!--DEV,PROD[,\s].*-->\s*(\n[\s\S]+?)^\s*<!-- -->\s*\n/gm,"$1\n");
+    // '${VERSION}'
+    const s3 = s.replace(/\${VERSION}/, version );
+    if (s3 !== s) return s3;
 
-  // PROD blocks (uncomment)
-  //
-  // If the '<!--PROD' has a tail, that is placed as a comment in the output.
-  //
-  const s3a = s2.replace(/^\s*<!--PROD\s*\n([\s\S]+?)^\s*-->\s*\n/gm, "\n$1\n");
-  const s3 = s3a.replace(/^(\s*)<!--PROD +(.*?)\s*\n([\s\S]+?)^\s*-->\s*\n/gm, "$1<!-- $2 -->\n$3\n" );
+    // Hashes
+    //
+    // Any '{file}-#.js' to be filled in the hash.
+    //
+    // Example:
+    //  <<
+    //    import { init } from './main-#.js'
+    //  <<
+    //
+    // Note: expects '-' delimiter in Rollup 'output.entryFileNames'
+    //
+    const s4 = s.replaceAll(/'[^']+\/([\w\d]+)-#\.js'/g,      // can there be multiple, per line?
+      (match,c1) => {   // e.g. match="'/main-#.js'", c1="main"
+        const hash = hashes.get(c1)
+        if (!hash) throw new Error( "No hash for: "+c1 );
 
-  // tbd. #rework: For the rest, might be better to go through them line by line, since they don't span multiple lines. (simpler regex)
+        return match.replace('#',hash);
+      }
+    );
+    return (s4 !== s) ? s4 : s;
+  });
 
-  // '${PRELOADS}'
-  //
-  // Note: A bit difficult playing with the white space and regex. '\s' matches also newlines. Is there a way to
-  //      "match white space but not newlines" other than '[ \t]'?
-  //
-  const s4 = s3.replace(/^([ \t]*)\${PRELOADS}[ \t]*$/gm, (_,c1) => preloads.map(x => c1+x).join('\n') );
-
-  // '${VERSION}'
-  //
-  const s5 = s4.replace(/\${VERSION}/gm, version );
-
-  // Hashes
-  //
-  // Any '{file}-#.js' to be filled in the hash.
-  //
-  // Example:
-  //  <<
-  //    import { init } from './main-#.js'
-  //  <<
-  //
-  // Note: expects '-' delimiter in Rollup 'output.entryFileNames'
-  //
-  const s6 = s5.replace(/'[^']+\/([\w\d]+)-#\.js'/gm,
-    (match,c1) => {   // e.g. match="'/main-#.js'", c1="main"
-      const hash = hashes.get(c1)
-      if (!hash) throw new Error( "No hash for: "+c1 );
-
-      return match.replace('#',hash);
-    }
-  );
-
-  return errors.length == 0 ? s6
-    : new Error( `ERROR(s) in generation of production 'index': ${errors.join()}` );
+  return outArr.join('\n');
 }
 
 /*
 * Rollup plugin to get the chunks and their hashes (as by the 'manualChunks' policy), pass it on to a tool script
-* that creates production index.html from the 'index.html' template.
+* that creates production 'index.html' from a template.
 *
 * Based on 'modulepreloadPlugin' by Philip Walton
 *   -> https://github.com/philipwalton/rollup-native-modules-boilerplate/blob/master/rollup.config.js#L63-L84