+++ packages/app-deploy-ops/src/assert.js
@@ -1,5 +1,5 @@
 /*
-* init/assert.js
+* src/assert.js
 *
 * Simple assert implementation.
 *