+++ packages/app-deploy-ops/src/central.js
@@ -1,5 +1,5 @@
 /*
-* init/central.js
+* src/central.js
 *
 * Central logging.
 *