+++ packages/app/CHANGELOG.md
@@ -4,6 +4,10 @@
 Release dates are marked; 'latest' shows WIP window.
 -->
 
+## latest
+
+- passing build version
+
 ## 6-Mar-2021
 
 - **FEATURE:** replaced FirebaseUI with `aside-keys` web component